+++ clang/lib/CMakeLists.txt
@@ -4,7 +4,6 @@ add_subdirectory(APINotes)
 add_subdirectory(Lex)
 add_subdirectory(Parse)
 add_subdirectory(AST)
-add_subdirectory(DepMatcher)
 add_subdirectory(ASTMatchers)
 add_subdirectory(CrossTU)
 add_subdirectory(Sema)
